requirements.txt
anthropic==0.94.0
instructor==1.15.1
tiktoken==0.12.0
pydantic==2.12.5
python-dotenv==1.2.2
streamlit==1.56.0
fastapi==0.135.3
uvicorn==0.44.0
httpx==0.28.1
python-multipart==0.0.26
pandas==3.0.2
matplotlib==3.10.8
faster-whisper==1.1.1
